Mirrored directory structure w/ install script
The `install.sh` script takes an input file and places it in the
location on your system as if the current working directory was your
root directory. This way files can be copied to the correct location
easily simply by treating this repo as if it were the root directory of
a file system. The `~` wildcard for the user's home directory is
recognized by the script.